Split locallib.php apart and use class autoloading
Moodle supports a class autoloading mechanism. Move the classes defined
in locallib.php into their own files under classes/.
