+++ strategy/strategy04_points.php
@@ -29,6 +29,8 @@
 // Namespace is mandatory!
 namespace mod_ratingallocate\strategy_points;
 
+use mod_ratingallocate\manager\strategymanager;
+
 defined('MOODLE_INTERNAL') || die();
 require_once($CFG->libdir . '/formslib.php');
 require_once(dirname(__FILE__) . '/../locallib.php');
@@ -128,7 +130,7 @@ class strategy extends \strategytemplate {
 }
 
 // Register with the strategymanager.
-\strategymanager::add_strategy(strategy::STRATEGYID);
+strategymanager::add_strategy(strategy::STRATEGYID);
 
 /**
  * View form.
