+++ strategy/strategy05_order.php
@@ -28,6 +28,8 @@
 // Namespace is mandatory!
 namespace mod_ratingallocate\strategy_order;
 
+use mod_ratingallocate\manager\strategymanager;
+
 defined('MOODLE_INTERNAL') || die();
 require_once($CFG->libdir . '/formslib.php');
 require_once(dirname(__FILE__) . '/../locallib.php');
@@ -113,7 +115,7 @@ class strategy extends \strategytemplate {
 }
 
 // Register with the strategymanager.
-\strategymanager::add_strategy(strategy::STRATEGYID);
+strategymanager::add_strategy(strategy::STRATEGYID);
 
 /**
  * _Users view_
