+++ strategy/strategy06_tickyes.php
@@ -29,6 +29,8 @@
 // Namespace is mandatory!
 namespace mod_ratingallocate\strategy_tickyes;
 
+use mod_ratingallocate\manager\strategymanager;
+
 defined('MOODLE_INTERNAL') || die();
 require_once($CFG->libdir . '/formslib.php');
 require_once(dirname(__FILE__) . '/../locallib.php');
@@ -123,7 +125,7 @@ class strategy extends \strategytemplate {
 }
 
 // Register with the strategymanager.
-\strategymanager::add_strategy(strategy::STRATEGYID);
+strategymanager::add_strategy(strategy::STRATEGYID);
 
 /**
  * _Users view_
