+++ tests/cron_test.php
@@ -31,11 +31,12 @@ use PHPUnit\Framework\Attributes\CoversClass;
  * Tests the correct behaviour of the cron task according to the rating period,
  * the current time and the current algorithm status.
  *
- * @package    mod_ratingallocate
- * @category   test
- * @group mod_ratingallocate
- * @copyright  T Reischmann 2015
- * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ * @package     mod_ratingallocate
+ * @category    test
+ * @group       mod_ratingallocate
+ * @copyright   T Reischmann 2015
+ * @license     http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ * @covers \mod_ratingallocate\task\cron_task
  */
 #[CoversClass(cron_task::class)]
 final class cron_test extends \advanced_testcase {
@@ -47,6 +48,9 @@ final class cron_test extends \advanced_testcase {
 
     /**
      * The cron should run, when the algorithm status is not_started and the rating period ended.
+     *
+     * @return void
+     * @covers \mod_ratingallocate\task\cron_task::execute
      */
     public function test_successful_run(): void {
         $this->create_ratingallocate(true);
@@ -56,6 +60,7 @@ final class cron_test extends \advanced_testcase {
 
     /**
      * The cron should not run, when the rating period has not ended.
+     * @covers \mod_ratingallocate\task\cron_task::execute
      */
     public function test_rating_in_progress(): void {
         $this->create_ratingallocate(false);
@@ -65,46 +70,51 @@ final class cron_test extends \advanced_testcase {
 
     /**
      * The cron should not run, when the algorithm status is running.
+     * @covers \mod_ratingallocate\task\cron_task::execute
      */
     public function test_running(): void {
-        $this->create_ratingallocate(true, \mod_ratingallocate\algorithm_status::RUNNING);
+        $this->create_ratingallocate(true, algorithm_status::RUNNING);
         $this->run_cron();
         $this->assert_running();
     }
 
     /**
      * The cron should not run, when the algorithm status is failure.
+     * @covers \mod_ratingallocate\task\cron_task::execute
      */
     public function test_failure(): void {
-        $this->create_ratingallocate(true, \mod_ratingallocate\algorithm_status::FAILURE);
+        $this->create_ratingallocate(true, algorithm_status::FAILURE);
         $this->run_cron();
         $this->assert_failure();
     }
 
     /**
      * The cron should not run, when the algorithm status is finished.
+     * @covers \mod_ratingallocate\task\cron_task::execute
      */
     public function test_finished(): void {
-        $this->create_ratingallocate(true, \mod_ratingallocate\algorithm_status::FINISHED);
+        $this->create_ratingallocate(true, algorithm_status::FINISHED);
         $this->run_cron();
         $this->assert_already_finish();
     }
 
     /**
      * The cron should not change the status of the algorithm, since it is not timedout.
+     * @covers \mod_ratingallocate\task\cron_task::execute
      */
     public function test_undue_failure_handling(): void {
-        $this->create_ratingallocate(true, \mod_ratingallocate\algorithm_status::RUNNING, time());
+        $this->create_ratingallocate(true, algorithm_status::RUNNING, time());
         $this->run_cron();
         $this->assert_running();
     }
 
     /**
      * The cron should switch the status of the algorithm to failure, since it has timedout.
+     * @covers \mod_ratingallocate\task\cron_task::execute
      */
     public function test_due_failure_handling(): void {
         global $CFG;
-        $this->create_ratingallocate(true, \mod_ratingallocate\algorithm_status::RUNNING, time() - 2);
+        $this->create_ratingallocate(true, algorithm_status::RUNNING, time() - 2);
         $CFG->ratingallocate_algorithm_timeout = 1;
         $this->run_cron();
         $this->assert_failure();
@@ -125,7 +135,7 @@ final class cron_test extends \advanced_testcase {
         global $DB;
         $record = $DB->get_record(this_db\ratingallocate::TABLE, []);
         $ratingallocate = \mod_ratingallocate_generator::get_ratingallocate_for_user($this, $record, $this->teacher);
-        $this->assertEquals(\mod_ratingallocate\algorithm_status::NOTSTARTED, $ratingallocate->get_algorithm_status());
+        $this->assertEquals(algorithm_status::NOTSTARTED, $ratingallocate->get_algorithm_status());
         $this->assertEquals(0, $DB->count_records(this_db\ratingallocate_allocations::TABLE,
                 [this_db\ratingallocate_allocations::RATINGALLOCATEID => $this->mod->id]));
     }
@@ -137,7 +147,7 @@ final class cron_test extends \advanced_testcase {
         global $DB;
         $record = $DB->get_record(this_db\ratingallocate::TABLE, []);
         $ratingallocate = \mod_ratingallocate_generator::get_ratingallocate_for_user($this, $record, $this->teacher);
-        $this->assertEquals(\mod_ratingallocate\algorithm_status::RUNNING, $ratingallocate->get_algorithm_status());
+        $this->assertEquals(algorithm_status::RUNNING, $ratingallocate->get_algorithm_status());
         $this->assertEquals(0, $DB->count_records(this_db\ratingallocate_allocations::TABLE,
                 [this_db\ratingallocate_allocations::RATINGALLOCATEID => $this->mod->id]));
     }
@@ -149,7 +159,7 @@ final class cron_test extends \advanced_testcase {
         global $DB;
         $record = $DB->get_record(this_db\ratingallocate::TABLE, []);
         $ratingallocate = \mod_ratingallocate_generator::get_ratingallocate_for_user($this, $record, $this->teacher);
-        $this->assertEquals(\mod_ratingallocate\algorithm_status::FAILURE, $ratingallocate->get_algorithm_status());
+        $this->assertEquals(algorithm_status::FAILURE, $ratingallocate->get_algorithm_status());
         $this->assertEquals(0, $DB->count_records(this_db\ratingallocate_allocations::TABLE,
                 [this_db\ratingallocate_allocations::RATINGALLOCATEID => $this->mod->id]));
     }
@@ -161,7 +171,7 @@ final class cron_test extends \advanced_testcase {
         global $DB;
         $record = $DB->get_record(this_db\ratingallocate::TABLE, []);
         $ratingallocate = \mod_ratingallocate_generator::get_ratingallocate_for_user($this, $record, $this->teacher);
-        $this->assertEquals(\mod_ratingallocate\algorithm_status::FINISHED, $ratingallocate->get_algorithm_status());
+        $this->assertEquals(algorithm_status::FINISHED, $ratingallocate->get_algorithm_status());
         $this->assertEquals(4, $DB->count_records(this_db\ratingallocate_allocations::TABLE,
                 [this_db\ratingallocate_allocations::RATINGALLOCATEID => $this->mod->id]));
     }
@@ -173,7 +183,7 @@ final class cron_test extends \advanced_testcase {
         global $DB;
         $record = $DB->get_record(this_db\ratingallocate::TABLE, []);
         $ratingallocate = \mod_ratingallocate_generator::get_ratingallocate_for_user($this, $record, $this->teacher);
-        $this->assertEquals(\mod_ratingallocate\algorithm_status::FINISHED, $ratingallocate->get_algorithm_status());
+        $this->assertEquals(algorithm_status::FINISHED, $ratingallocate->get_algorithm_status());
         $this->assertEquals(0, $DB->count_records(this_db\ratingallocate_allocations::TABLE,
                 [this_db\ratingallocate_allocations::RATINGALLOCATEID => $this->mod->id]));
     }
@@ -185,7 +195,7 @@ final class cron_test extends \advanced_testcase {
      * @param int $algorithmstarttime the start time of the algorithm.
      */
     private function create_ratingallocate($ratingperiodended,
-            $algorithmstatus = \mod_ratingallocate\algorithm_status::NOTSTARTED, $algorithmstarttime = null) {
+                                           $algorithmstatus = algorithm_status::NOTSTARTED, $algorithmstarttime = null) {
         global $DB;
 
         $this->resetAfterTest();
